+++ saleor/graphql/meta/tests/test_meta_mutations.py
@@ -1,5 +1,6 @@
 import base64
 import uuid
+import warnings
 from unittest.mock import patch
 
 import graphene
@@ -341,6 +342,21 @@ def test_add_public_metadata_for_checkout(api_client, checkout):
     )
 
 
+def test_add_public_metadata_for_checkout_by_token(api_client, checkout):
+    # given
+    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
+
+    # when
+    response = execute_update_public_metadata_for_item(
+        api_client, None, checkout.token, "Checkout"
+    )
+
+    # then
+    assert item_contains_proper_public_metadata(
+        response["data"]["updateMetadata"]["item"], checkout, checkout_id
+    )
+
+
 @patch("saleor.plugins.manager.PluginsManager.checkout_updated")
 def test_add_metadata_for_checkout_triggers_checkout_updated_hook(
     mock_checkout_updated, api_client, checkout
@@ -353,13 +369,32 @@ def test_add_metadata_for_checkout_triggers_checkout_updated_hook(
     mock_checkout_updated.assert_called_once_with(checkout)
 
 
-def test_add_public_metadata_for_order(api_client, order):
+def test_add_public_metadata_for_order_by_id(api_client, order):
+    # given
+    order_id = graphene.Node.to_global_id("Order", order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_update_public_metadata_for_item(
+            api_client, None, order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_contains_proper_public_metadata(
+        response["data"]["updateMetadata"]["item"], order, order_id
+    )
+
+
+def test_add_public_metadata_for_order_by_token(api_client, order):
     # given
     order_id = graphene.Node.to_global_id("Order", order.pk)
 
     # when
     response = execute_update_public_metadata_for_item(
-        api_client, None, order_id, "Order"
+        api_client, None, order.token, "Order"
     )
 
     # then
@@ -368,13 +403,32 @@ def test_add_public_metadata_for_order(api_client, order):
     )
 
 
-def test_add_public_metadata_for_draft_order(api_client, draft_order):
+def test_add_public_metadata_for_draft_order_by_id(api_client, draft_order):
+    # given
+    draft_order_id = graphene.Node.to_global_id("Order", draft_order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_update_public_metadata_for_item(
+            api_client, None, draft_order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_contains_proper_public_metadata(
+        response["data"]["updateMetadata"]["item"], draft_order, draft_order_id
+    )
+
+
+def test_add_public_metadata_for_draft_order_by_token(api_client, draft_order):
     # given
     draft_order_id = graphene.Node.to_global_id("Order", draft_order.pk)
 
     # when
     response = execute_update_public_metadata_for_item(
-        api_client, None, draft_order_id, "Order"
+        api_client, None, draft_order.token, "Order"
     )
 
     # then
@@ -1080,7 +1134,45 @@ def test_delete_public_metadata_for_checkout(api_client, checkout):
     )
 
 
-def test_delete_public_metadata_for_order(api_client, order):
+def test_delete_public_metadata_for_checkout_by_token(api_client, checkout):
+    # given
+    checkout.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    checkout.save(update_fields=["metadata"])
+    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
+
+    # when
+    response = execute_clear_public_metadata_for_item(
+        api_client, None, checkout.token, "Checkout"
+    )
+
+    # then
+    assert item_without_public_metadata(
+        response["data"]["deleteMetadata"]["item"], checkout, checkout_id
+    )
+
+
+def test_delete_public_metadata_for_order_by_id(api_client, order):
+    # given
+    order.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    order.save(update_fields=["metadata"])
+    order_id = graphene.Node.to_global_id("Order", order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_clear_public_metadata_for_item(
+            api_client, None, order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_without_public_metadata(
+        response["data"]["deleteMetadata"]["item"], order, order_id
+    )
+
+
+def test_delete_public_metadata_for_order_by_token(api_client, order):
     # given
     order.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
     order.save(update_fields=["metadata"])
@@ -1088,7 +1180,7 @@ def test_delete_public_metadata_for_order(api_client, order):
 
     # when
     response = execute_clear_public_metadata_for_item(
-        api_client, None, order_id, "Order"
+        api_client, None, order.token, "Order"
     )
 
     # then
@@ -1097,7 +1189,28 @@ def test_delete_public_metadata_for_order(api_client, order):
     )
 
 
-def test_delete_public_metadata_for_draft_order(api_client, draft_order):
+def test_delete_public_metadata_for_draft_order_by_id(api_client, draft_order):
+    # given
+    draft_order.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
+    draft_order.save(update_fields=["metadata"])
+    draft_order_id = graphene.Node.to_global_id("Order", draft_order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_clear_public_metadata_for_item(
+            api_client, None, draft_order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_without_public_metadata(
+        response["data"]["deleteMetadata"]["item"], draft_order, draft_order_id
+    )
+
+
+def test_delete_public_metadata_for_draft_order_by_token(api_client, draft_order):
     # given
     draft_order.store_value_in_metadata({PUBLIC_KEY: PUBLIC_VALUE})
     draft_order.save(update_fields=["metadata"])
@@ -1105,7 +1218,7 @@ def test_delete_public_metadata_for_draft_order(api_client, draft_order):
 
     # when
     response = execute_clear_public_metadata_for_item(
-        api_client, None, draft_order_id, "Order"
+        api_client, None, draft_order.token, "Order"
     )
 
     # then
@@ -1834,7 +1947,45 @@ def test_add_private_metadata_for_checkout(
     )
 
 
-def test_add_private_metadata_for_order(
+def test_add_private_metadata_for_checkout_by_token(
+    staff_api_client, checkout, permission_manage_checkouts
+):
+    # given
+    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
+
+    # when
+    response = execute_update_private_metadata_for_item(
+        staff_api_client, permission_manage_checkouts, checkout.token, "Checkout"
+    )
+
+    # then
+    assert item_contains_proper_private_metadata(
+        response["data"]["updatePrivateMetadata"]["item"], checkout, checkout_id
+    )
+
+
+def test_add_private_metadata_for_order_by_id(
+    staff_api_client, order, permission_manage_orders
+):
+    # given
+    order_id = graphene.Node.to_global_id("Order", order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_update_private_metadata_for_item(
+            staff_api_client, permission_manage_orders, order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_contains_proper_private_metadata(
+        response["data"]["updatePrivateMetadata"]["item"], order, order_id
+    )
+
+
+def test_add_private_metadata_for_order_by_token(
     staff_api_client, order, permission_manage_orders
 ):
     # given
@@ -1842,7 +1993,7 @@ def test_add_private_metadata_for_order(
 
     # when
     response = execute_update_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, order_id, "Order"
+        staff_api_client, permission_manage_orders, order.token, "Order"
     )
 
     # then
@@ -1851,7 +2002,28 @@ def test_add_private_metadata_for_order(
     )
 
 
-def test_add_private_metadata_for_draft_order(
+def test_add_private_metadata_for_draft_order_by_id(
+    staff_api_client, draft_order, permission_manage_orders
+):
+    # given
+    draft_order_id = graphene.Node.to_global_id("Order", draft_order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_update_private_metadata_for_item(
+            staff_api_client, permission_manage_orders, draft_order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_contains_proper_private_metadata(
+        response["data"]["updatePrivateMetadata"]["item"], draft_order, draft_order_id
+    )
+
+
+def test_add_private_metadata_for_draft_order_by_token(
     staff_api_client, draft_order, permission_manage_orders
 ):
     # given
@@ -1859,7 +2031,7 @@ def test_add_private_metadata_for_draft_order(
 
     # when
     response = execute_update_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, draft_order_id, "Order"
+        staff_api_client, permission_manage_orders, draft_order.token, "Order"
     )
 
     # then
@@ -2630,7 +2802,49 @@ def test_delete_private_metadata_for_checkout(
     )
 
 
-def test_delete_private_metadata_for_order(
+def test_delete_private_metadata_for_checkout_by_token(
+    staff_api_client, checkout, permission_manage_checkouts
+):
+    # given
+    checkout.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
+    checkout.save(update_fields=["private_metadata"])
+    checkout_id = graphene.Node.to_global_id("Checkout", checkout.pk)
+
+    # when
+    response = execute_clear_private_metadata_for_item(
+        staff_api_client, permission_manage_checkouts, checkout.token, "Checkout"
+    )
+
+    # then
+    assert item_without_private_metadata(
+        response["data"]["deletePrivateMetadata"]["item"], checkout, checkout_id
+    )
+
+
+def test_delete_private_metadata_for_order_by_id(
+    staff_api_client, order, permission_manage_orders
+):
+    # given
+    order.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
+    order.save(update_fields=["private_metadata"])
+    order_id = graphene.Node.to_global_id("Order", order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_clear_private_metadata_for_item(
+            staff_api_client, permission_manage_orders, order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_without_private_metadata(
+        response["data"]["deletePrivateMetadata"]["item"], order, order_id
+    )
+
+
+def test_delete_private_metadata_for_order_by_token(
     staff_api_client, order, permission_manage_orders
 ):
     # given
@@ -2640,7 +2854,7 @@ def test_delete_private_metadata_for_order(
 
     # when
     response = execute_clear_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, order_id, "Order"
+        staff_api_client, permission_manage_orders, order.token, "Order"
     )
 
     # then
@@ -2649,7 +2863,30 @@ def test_delete_private_metadata_for_order(
     )
 
 
-def test_delete_private_metadata_for_draft_order(
+def test_delete_private_metadata_for_draft_order_by_id(
+    staff_api_client, draft_order, permission_manage_orders
+):
+    # given
+    draft_order.store_value_in_private_metadata({PRIVATE_KEY: PRIVATE_VALUE})
+    draft_order.save(update_fields=["private_metadata"])
+    draft_order_id = graphene.Node.to_global_id("Order", draft_order.pk)
+
+    # when
+    with warnings.catch_warnings(record=True) as warns:
+        response = execute_clear_private_metadata_for_item(
+            staff_api_client, permission_manage_orders, draft_order_id, "Order"
+        )
+        expected_warning = "DEPRECATED. Use token for changing order metadata."
+
+        assert any([str(warning.message) == expected_warning for warning in warns])
+
+    # then
+    assert item_without_private_metadata(
+        response["data"]["deletePrivateMetadata"]["item"], draft_order, draft_order_id
+    )
+
+
+def test_delete_private_metadata_for_draft_order_by_token(
     staff_api_client, draft_order, permission_manage_orders
 ):
     # given
@@ -2659,7 +2896,7 @@ def test_delete_private_metadata_for_draft_order(
 
     # when
     response = execute_clear_private_metadata_for_item(
-        staff_api_client, permission_manage_orders, draft_order_id, "Order"
+        staff_api_client, permission_manage_orders, draft_order.token, "Order"
     )
 
     # then